requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-extra-fields==3.0.2
django-filter==25.1
django-markdownx==4.0.7
django-mdeditor==0.1.20
django-soft-delete==1.0.18
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
Markdown==3.7
markdownify==1.1.0
pillow==11.1.0
# psycopg2==2.9.10
# psycopg2-binary==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
python-magic==0.4.27
python-magic-bin==0.4.14
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
